feat(update): enhance user profile and registration flow with phone and country data
- Increase referral and ad reward amounts from 0.5 to 2 EKH
- Add name field to UserProfile data model
- Refactor leaderboard queries to use totalCoins field exclusively
- Fix miningReward type to Long in repositories and database updates
- Add UpdateCheckWrapper composable to main UI for app updates
- Implement EditProfileScreen with username and full name editing capability
- Remove obsolete EditUsernameScreen UI elements for profile picture
- Extend EmailSignupScreen with phone number and country selection fields
- Display warning about importance of phone number for password recovery
- Include CountryDropdownField for country selection in signup and OAuth completion
- Add GoogleSignInButton composable and replace inline buttons with it
- Add FileProvider in AndroidManifest for APK installation support
- Increase app versionCode to 10 and update versionName to 1.0.0
- Adjust UI text color of some buttons and texts for better visibility
- Reduce font sizes in MiningScreen user profile card for layout improvement

+++ ktMobileApp/app/src/main/java/com/ekehi/network/data/repository/MiningRepository.kt
@@ -142,7 +142,7 @@ open class MiningRepository @Inject constructor(
                         collectionId = AppwriteService.USER_PROFILES_COLLECTION,
                         documentId = profile.id,
                         data = mapOf(
-                                "miningReward" to (currentMiningReward + reward),
+                                "miningReward" to (currentMiningReward + reward).toLong(),
                                 "todayEarnings" to (todayEarnings + reward),
                                 "updatedAt" to currentDate
                         )
